Improved rnd_requirements and min_requirements
Allow installing . for `pyexcel` in rnd_requirements,
which fails with setuptoos 17.1 so that is upgraded to the
latest version.

Also add a MINREQ mode with a min_requirements.txt template
to generate a requirements.txt that requires the minimum
version instead of allowing pip to install the latest version,
to ensure that the minimum requirements are in fact still
supported.

And workaround the fact that texttable 0.8.1 and 0.8.2
are not located in pypi or github.